I required a new include to fill a campaign dropdown in the user administration
Users can now also be assigned/removed to/from campaigns by the
administrator
deleting a user also deletes the entries inthe campaign_users table
aligned variable for user_id (formerly id) with table

+++ Website/UserManagementModify.php
@@ -28,13 +28,18 @@
 				
 				# bind post variable into variables
 				#  User id
-					$id = $_POST["id"];
+					$id = $_POST["userid"];			
 				# encrypt pasword
 					$password = md5($_POST["password"]);
 				# User role
-					$role = $_POST["role"];		
+					$role = $_POST["role"];	
+				# campaign database
+					$campdb = $_POST["campdb"];
+				# campaign database for removal
+					$remCampdb = $_POST["remCampdb"];					
+
 				# get user name
-					$sql = "SELECT username FROM users WHERE id = $id";
+					$sql = "SELECT username FROM users WHERE user_id = $id";
 					
 					if(!$result = $dbc->query($sql)){
 					die('There was an error running the query ' . mysqli_error($dbc));
@@ -42,45 +47,62 @@
 					# load the name into a variable 
 					while($row = $result->fetch_assoc()) 
 					 {
-					$username = $row['username'];
+					$modifieduser = $row['username'];
 					 }
-			
 				# If a users should be deleted
 				if (($_POST["modify"] == 0))
 					{
-						$sql = "DELETE FROM users WHERE	id = '$id'";
+						$sql = "DELETE FROM users WHERE	user_id = '$id'; DELETE FROM campaign_users where user_id = '$id'";
 					}
 					else
 				# if a user wants to have his password reset
 				if (($_POST["modify"] == 1))
 					{
-						$sql = "UPDATE users set password = '$password' WHERE id = '$id'";
+						$sql = "UPDATE users set password = '$password' WHERE user_id = '$id'";
 					}
 					else
 				# if a user wants to have another role
 				if (($_POST["modify"] == 2))
 					{
-						$sql = "UPDATE users set role = \"$role\" WHERE id = '$id'";
+						$sql = "UPDATE users set role = \"$role\" WHERE user_id = '$id'";
 					}
+				# if a user is assigned to a campaign
+				if (($_POST["modify"] == 3))
+					{	
+						$sql = "INSERT into campaign_users (user_id, camp_db, CoalID) VALUES ('$id', '$campdb', 0)";
+					}
+				# remove a user from a campaign
+				if (($_POST["modify"] == 4))
+					{	
+						$sql = "DELETE FROM campaign_users where user_id = '$id' and camp_db = '$remCampdb'";
+					}										
 				
 				# Feedback success or failure
-				if (!mysqli_query($dbc,$sql))
+				if (!mysqli_multi_query($dbc,$sql))
 					{
 						die('<br>There was an error running the query: ' . mysqli_error($dbc));
 					}
 				  
 				if (($_POST["modify"] == 0)) 
 					{
-						echo "<br>Record <b>$id</b> owned by user <b>$username</b> deleted successfully!\n";
+						echo "<br>Record <b>$id</b> owned by user <b>$modifieduser</b> deleted successfully!\n";
 					}
 				if (($_POST["modify"] == 1))
 					{
-						echo "<br>Password for user record <b>$id</b> owned by user <b>$username</b> updated successfully!\n";
+						echo "<br>Password for user record <b>$id</b> owned by user <b>$modifieduser</b> updated successfully!\n";
 					}
 				if (($_POST["modify"] == 2))
 					{
-						echo "<br>Role for user <b>$id</b> owned by user <b>$username</b> updated successfully to <b>$role</b>!\n";
+						echo "<br>Role for user <b>$id</b> owned by user <b>$modifieduser</b> updated successfully to <b>$role</b>!\n";
 					}
+				if (($_POST["modify"] == 3))
+					{
+						echo "<br>The user <b>$id</b> has been assigned to the <b>$campdb</b> campaign successfully!\n";
+					}
+				if (($_POST["modify"] == 4))
+					{
+						echo "<br>The user <b>$id</b> has been removed from the <b>$campdb</b> campaign successfully!\n";
+					}										
 			
 				?>
